api/requirements.txt
tensorflow==2.10.0
numpy==1.24.4
Pillow
gradio
huggingface_hub
fastapi
uvicorn
opencv-python
inference-sdk